Update product pricing handling to use decimal values for USD prices
- Changed price handling in EditProductPage, NewProductPage, and API route to use parseFloat for USD prices instead of parseInt, allowing for decimal values.
- Updated input fields to reflect pricing in dollars instead of cents, including adjustments to minimum values and placeholders.
- Modified translation strings to clarify pricing format in both English and Spanish.

+++ src/app/api/admin/products/route.ts
@@ -22,9 +22,9 @@ export async function POST(request: NextRequest) {
     const slug = formData.get("slug") as string;
     const description = formData.get("description") as string;
     const image = formData.get("image") as string;
-    const priceUSD = parseInt(formData.get("priceUSD") as string);
+    const priceUSD = parseFloat(formData.get("priceUSD") as string);
     const priceCLP = parseInt(formData.get("priceCLP") as string);
-    const salePriceUSD = formData.get("salePriceUSD") ? parseInt(formData.get("salePriceUSD") as string) : null;
+    const salePriceUSD = formData.get("salePriceUSD") ? parseFloat(formData.get("salePriceUSD") as string) : null;
     const salePriceCLP = formData.get("salePriceCLP") ? parseInt(formData.get("salePriceCLP") as string) : null;
     const defaultDurationDays = parseInt(formData.get("defaultDurationDays") as string);
     const maxActivations = parseInt(formData.get("maxActivations") as string);
